extract nexus file reader into a separate directory to be used by consumer as well as producer
